# docs/swagger.yaml
basePath: /api
definitions:
  controllers.CreateCommentInput:
    properties:
      content:
        type: string
      parent_id:
        type: integer
      post_id:
        type: integer
    required:
    - content
    - post_id
    type: object
  controllers.CreatePostInput:
    properties:
      content:
        type: string
      is_draft:
        type: boolean
      is_recommend:
        type: boolean
      is_top:
        type: boolean
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    required:
    - content
    - title
    type: object
  controllers.LikeInput:
    properties:
      target_id:
        type: integer
      target_type:
        enum:
        - post
        - comment
        type: string
    required:
    - target_id
    - target_type
    type: object
  controllers.UpdatePostInput:
    properties:
      content:
        type: string
      is_draft:
        type: boolean
      is_recommend:
        type: boolean
      is_top:
        type: boolean
      title:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 展示接口功能
  title: GoBlog API文档
  version: "1.0"
paths:
  /comments:
    get:
      consumes:
      - application/json
      parameters:
      - description: 文章 ID
        in: query
        name: post_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 获取文章下的评论列表（带子评论）
      tags:
      - 评论
    post:
      consumes:
      - application/json
      parameters:
      - description: 评论数据
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateCommentInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: 创建评论或回复
      tags:
      - 评论
  /likes:
    post:
      consumes:
      - application/json
      parameters:
      - description: 点赞对象
        in: body
        name: like
        required: true
        schema:
          $ref: '#/definitions/controllers.LikeInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: 点赞文章或评论
      tags:
      - 点赞
  /likes/check:
    get:
      consumes:
      - application/json
      parameters:
      - description: 目标 ID
        in: query
        name: target_id
        required: true
        type: integer
      - description: 目标类型（post/comment）
        in: query
        name: target_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: 检查用户是否已点赞
      tags:
      - 点赞
  /likes/count:
    get:
      consumes:
      - application/json
      parameters:
      - description: 目标 ID
        in: query
        name: target_id
        required: true
        type: integer
      - description: 目标类型（post/comment）
        in: query
        name: target_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
      summary: 获取点赞总数
      tags:
      - 点赞
  /login:
    post:
      consumes:
      - application/json
      description: 使用用户名或邮箱+密码登录，返回 JWT Token
      parameters:
      - description: 用户名和密码
        in: body
        name: credentials
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 用户登录
      tags:
      - 用户
  /posts:
    get:
      consumes:
      - application/json
      description: 支持分页获取文章列表
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 获取文章列表
      tags:
      - 文章
    post:
      consumes:
      - application/json
      description: 登录用户可发布新文章，并附带标签
      parameters:
      - description: 文章数据
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/controllers.CreatePostInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: 创建文章
      tags:
      - 文章
  /posts/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 文章 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: 删除文章
      tags:
      - 文章
    get:
      consumes:
      - application/json
      parameters:
      - description: 文章 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取文章详情
      tags:
      - 文章
    put:
      consumes:
      - application/json
      parameters:
      - description: 文章 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新数据
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdatePostInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: 修改文章
      tags:
      - 文章
  /register:
    post:
      consumes:
      - application/json
      description: 用户通过用户名、邮箱、密码注册账号
      parameters:
      - description: 注册信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 用户注册
      tags:
      - 用户
  /tags/{name}/posts:
    get:
      consumes:
      - application/json
      parameters:
      - description: 标签名
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 获取指定标签下的所有文章
      tags:
      - 标签
swagger: "2.0"